refactor: split canvas page model into focused hooks
