+++ frontend/src/pages/canvas/model/useCanvasPage.ts
@@ -1,69 +1,14 @@
-import { useCallback, useEffect, useRef, useState } from "react";
+import { useCallback, useEffect, useRef } from "react";
 import {
   useVotePopup,
   useVoteState,
-  useVoteTickets,
 } from "@/features/gameplay/vote";
-import {
-  Cell,
-  CELL_SIZE,
-  useCanvasInteraction,
-  useCanvasNavigation,
-  useCanvasRenderer,
-  useCanvasViewport,
-} from "@/features/gameplay/canvas";
-import {
-  useRoundState,
-  useRoundTimer,
-} from "@/features/gameplay/round";
-import {
-  SessionBootstrapResult,
-  useGameSession,
-  useGameplaySocket,
-} from "@/features/gameplay/session";
-
-function formatClockTime(date: Date): string {
-  const hours = String(date.getHours()).padStart(2, "0");
-  const minutes = String(date.getMinutes()).padStart(2, "0");
-  return `${hours}:${minutes}`;
-}
+import type { SessionBootstrapResult } from "@/features/gameplay/session";
+import useCanvasGameplay from "./useCanvasGameplay";
+import useCanvasScene from "./useCanvasScene";
 
 export default function useCanvasPage() {
-  const canvasRef = useRef<HTMLCanvasElement>(null);
-  const containerRef = useRef<HTMLDivElement>(null);
-
-  const cellsRef = useRef<Cell[]>([]);
-  const selectedCellRef = useRef<Cell | null>(null);
   const isRoundExpiredRef = useRef(false);
-
-  const [cells, setCells] = useState<Cell[]>([]);
-  const [canvasId, setCanvasId] = useState<number | null>(null);
-  const [gridX, setGridX] = useState(0);
-  const [gridY, setGridY] = useState(0);
-  const [selectedCell, setSelectedCell] = useState<Cell | null>(null);
-  const [canvasReady, setCanvasReady] = useState(false);
-
-  const {
-    roundId,
-    roundNumber,
-    roundDurationSec,
-    totalRounds,
-    formattedGameEndTime,
-    setRoundState,
-    applyRoundMeta,
-    resetRoundState,
-  } = useRoundState();
-
-  const {
-    remainingSeconds,
-    formattedRemainingTime,
-    isRoundExpired,
-    setRoundTimerState,
-    applyRoundTimer,
-    startRoundTimer,
-    expireRoundTimer,
-    resetRoundTimer,
-  } = useRoundTimer();
 
   const {
     popupOpen,
@@ -83,259 +28,74 @@ export default function useCanvasPage() {
     resetVoteState,
   } = useVoteState();
 
-  const { remaining, setRemaining, fetchTickets, clearTickets } =
-    useVoteTickets();
+  const {
+    canvasRef,
+    containerRef,
+    cells,
+    canvasId,
+    gridX,
+    gridY,
+    selectedCell,
+    viewport,
+    navigateToCoordinate,
+    setCanvasId,
+    setGridX,
+    setGridY,
+    updateCells,
+    handleMouseDown,
+    handleMouseMove,
+    handleMouseUp,
+    handleMouseLeave,
+    handleCanvasUpdated,
+    clearSelectedCell,
+  } = useCanvasScene({
+    previewColorRef,
+    votingCellIdsRef,
+    topColorMapRef,
+    isRoundExpiredRef,
+    resetPreviewColor,
+    openPopup,
+    closePopup,
+  });
 
-  useEffect(() => {
-    selectedCellRef.current = selectedCell;
-  }, [selectedCell]);
-
-  useEffect(() => {
-    isRoundExpiredRef.current = isRoundExpired;
-  }, [isRoundExpired]);
-
-  const updateCells = useCallback(
-    (updater: Cell[] | ((prev: Cell[]) => Cell[])) => {
-      if (typeof updater === "function") {
-        setCells((prev) => {
-          const next = updater(prev);
-          cellsRef.current = next;
-          return next;
-        });
-      } else {
-        cellsRef.current = updater;
-        setCells(updater);
-      }
-    },
-    [],
-  );
-
-  const applyBootstrap = useCallback(
+  const applyBootstrapScene = useCallback(
     (result: SessionBootstrapResult) => {
       setCanvasId(result.canvasId);
       setGridX(result.gridX);
       setGridY(result.gridY);
       updateCells(result.cells);
-
-      setRoundState({
-        roundId: result.round.roundId,
-        roundNumber: result.round.roundNumber,
-        roundDurationSec: result.round.roundDurationSec,
-        totalRounds: result.round.totalRounds,
-        formattedGameEndTime: result.round.formattedGameEndTime,
-      });
-
-      setRoundTimerState({
-        remainingSeconds: result.round.remainingSeconds,
-        formattedRemainingTime: result.round.formattedRemainingTime,
-        isRoundExpired: result.round.isRoundExpired,
-      });
-
-      setRemaining(result.remaining);
     },
-    [setRemaining, setRoundState, setRoundTimerState, updateCells],
+    [setCanvasId, setGridX, setGridY, updateCells],
   );
 
-  const {
-    loading,
-    error,
-    gameEnded,
-    initializeSession,
-    clearSessionError,
-    markGameEnded,
-  } = useGameSession({
-    onBootstrap: applyBootstrap,
-  });
+  const handleGameEndedCleanup = useCallback(() => {
+    clearSelectedCell();
+  }, [clearSelectedCell]);
 
-  useEffect(() => {
-    initializeSession();
-  }, [initializeSession]);
-
-  useEffect(() => {
-    const canvas = canvasRef.current;
-    if (!canvas || !canvasId || gridX === 0 || gridY === 0) return;
-
-    canvas.width = gridX * CELL_SIZE;
-    canvas.height = gridY * CELL_SIZE;
-    setCanvasReady(true);
-  }, [canvasId, gridX, gridY]);
-
-  const { viewport, updateViewport } = useCanvasViewport({
-    containerRef,
-    canvasRef,
-    gridX,
-    gridY,
-    canvasReady,
-  });
-
-  const { navigateToCoordinate } = useCanvasNavigation({
-    containerRef,
-    canvasRef,
-    updateViewport,
-  });
-
-  useCanvasRenderer({
-    canvasRef,
-    canvasReady,
-    cellsRef,
-    selectedCellRef,
-    previewColorRef,
-    votingCellIdsRef,
-    topColorMapRef,
-  });
-
-  const handleSelectCell = useCallback((cell: Cell) => {
-    setSelectedCell(cell);
-    selectedCellRef.current = cell;
-  }, []);
-
-  const { handleMouseDown, handleMouseMove, handleMouseUp, handleMouseLeave } =
-    useCanvasInteraction({
-      containerRef,
-      canvasRef,
-      cells,
-      onSelectCell: handleSelectCell,
-      onResetPreviewColor: resetPreviewColor,
-      onOpenPopup: openPopup,
-    });
-
-  const handleRoundStarted = useCallback(
-    async ({
-      roundId,
-      roundNumber,
-      totalRounds,
-      gameEndAt,
-      roundDurationSec,
-    }: {
-      roundId: number;
-      roundNumber: number;
-      startedAt: string;
-      roundDurationSec: number;
-      totalRounds: number;
-      gameEndAt: string;
-    }) => {
-      setRoundState({
-        roundId,
-        roundNumber,
-        roundDurationSec,
-        totalRounds,
-        formattedGameEndTime: formatClockTime(new Date(gameEndAt)),
-      });
-
-      startRoundTimer(roundDurationSec);
-      clearSessionError();
-      resetVoteState();
-      await fetchTickets(roundId);
-    },
-    [
-      clearSessionError,
-      fetchTickets,
-      resetVoteState,
-      setRoundState,
-      startRoundTimer,
-    ],
-  );
-
-  const handleRoundEnded = useCallback(() => {
-    clearTickets();
-    expireRoundTimer();
-    resetVoteState();
-  }, [clearTickets, expireRoundTimer, resetVoteState]);
-
-  const handleCanvasUpdated = useCallback(
-    ({ cellId, color }: { cellId: number; color: string }) => {
-      updateCells((prev) =>
-        prev.map((cell) =>
-          cell.id === cellId ? { ...cell, color, status: "painted" } : cell,
-        ),
-      );
-
-      if (selectedCellRef.current?.id === cellId) {
-        const nextSelectedCell: Cell = {
-          ...selectedCellRef.current,
-          color,
-          status: "painted",
-        };
-        setSelectedCell(nextSelectedCell);
-        selectedCellRef.current = nextSelectedCell;
-
-        if (!isRoundExpiredRef.current) {
-          closePopup();
-        }
-      }
-    },
-    [closePopup, updateCells],
-  );
-
-  const handleVoteUpdate = useCallback(
-    ({ votes }: { votes: Record<string, number> }) => {
-      applyVoteUpdate(votes);
-    },
-    [applyVoteUpdate],
-  );
-
-  const handleTimerUpdate = useCallback(
-    ({
-      remainingSeconds,
-      isRoundExpired,
-      gameEndAt,
-      roundDurationSec,
-      totalRounds,
-    }: {
-      remainingSeconds: number;
-      isRoundExpired: boolean;
-      gameEndAt: string;
-      roundDurationSec: number;
-      totalRounds: number;
-    }) => {
-      applyRoundTimer({ remainingSeconds, isRoundExpired });
-      applyRoundMeta({ roundDurationSec, totalRounds, gameEndAt });
-    },
-    [applyRoundMeta, applyRoundTimer],
-  );
-
-  const handleGameEnded = useCallback(() => {
-    markGameEnded();
-    clearTickets();
-    resetRoundState();
-    resetRoundTimer();
-    resetVoteState();
-  }, [
-    clearTickets,
-    markGameEnded,
-    resetRoundState,
-    resetRoundTimer,
-    resetVoteState,
-  ]);
-
-  useGameplaySocket({
+  const gameplay = useCanvasGameplay({
     canvasId,
-    onRoundStarted: handleRoundStarted,
-    onRoundEnded: handleRoundEnded,
+    onBootstrapScene: applyBootstrapScene,
     onCanvasUpdated: handleCanvasUpdated,
-    onVoteUpdate: handleVoteUpdate,
-    onTimerUpdate: handleTimerUpdate,
-    onGameEnded: handleGameEnded,
+    onGameEndedCleanup: handleGameEndedCleanup,
+    applyVoteUpdate,
+    resetVoteState,
   });
 
-  const handleVoteSuccess = useCallback(() => {
-    if (roundId) {
-      void fetchTickets(roundId);
-    }
-  }, [fetchTickets, roundId]);
+  useEffect(() => {
+    isRoundExpiredRef.current = gameplay.isRoundExpiredRefValue;
+  }, [gameplay.isRoundExpiredRefValue]);
 
   const handlePopupClose = useCallback(() => {
-    setSelectedCell(null);
-    selectedCellRef.current = null;
+    clearSelectedCell();
     closePopup();
-  }, [closePopup]);
+  }, [clearSelectedCell, closePopup]);
 
   return {
     canvasRef,
     containerRef,
-    loading,
-    error,
-    gameEnded,
+    loading: gameplay.loading,
+    error: gameplay.error,
+    gameEnded: gameplay.gameEnded,
     handleMouseDown,
     handleMouseMove,
     handleMouseUp,
@@ -343,21 +103,21 @@ export default function useCanvasPage() {
     popupOpen,
     popupPos,
     canvasId,
-    roundId,
-    isRoundExpired,
+    roundId: gameplay.roundId,
+    isRoundExpired: gameplay.isRoundExpired,
     selectedCell,
     votes,
     cells,
-    handleVoteSuccess,
+    handleVoteSuccess: gameplay.handleVoteSuccess,
     handleColorChange,
     handlePopupClose,
-    roundNumber,
-    totalRounds,
-    formattedGameEndTime,
-    formattedRemainingTime,
-    remainingSeconds,
-    roundDurationSec,
-    remaining,
+    roundNumber: gameplay.roundNumber,
+    totalRounds: gameplay.totalRounds,
+    formattedGameEndTime: gameplay.formattedGameEndTime,
+    formattedRemainingTime: gameplay.formattedRemainingTime,
+    remainingSeconds: gameplay.remainingSeconds,
+    roundDurationSec: gameplay.roundDurationSec,
+    remaining: gameplay.remaining,
     gridX,
     gridY,
     viewport,
