+++ frontend/src/pages/canvas/model/useCanvasScene.ts
@@ -1,0 +1,165 @@
+import { useCallback, useEffect, useRef, useState } from "react";
+import {
+  Cell,
+  CELL_SIZE,
+  useCanvasInteraction,
+  useCanvasNavigation,
+  useCanvasRenderer,
+  useCanvasViewport,
+} from "@/features/gameplay/canvas";
+
+interface UseCanvasSceneParams {
+  previewColorRef: React.RefObject<string | null>;
+  votingCellIdsRef: React.RefObject<Set<number>>;
+  topColorMapRef: React.RefObject<Map<number, string>>;
+  isRoundExpiredRef: React.RefObject<boolean>;
+  resetPreviewColor: () => void;
+  openPopup: (position: { x: number; y: number }) => void;
+  closePopup: () => void;
+}
+
+export default function useCanvasScene({
+  previewColorRef,
+  votingCellIdsRef,
+  topColorMapRef,
+  isRoundExpiredRef,
+  resetPreviewColor,
+  openPopup,
+  closePopup,
+}: UseCanvasSceneParams) {
+  const canvasRef = useRef<HTMLCanvasElement>(null);
+  const containerRef = useRef<HTMLDivElement>(null);
+
+  const cellsRef = useRef<Cell[]>([]);
+  const selectedCellRef = useRef<Cell | null>(null);
+
+  const [cells, setCells] = useState<Cell[]>([]);
+  const [canvasId, setCanvasId] = useState<number | null>(null);
+  const [gridX, setGridX] = useState(0);
+  const [gridY, setGridY] = useState(0);
+  const [selectedCell, setSelectedCell] = useState<Cell | null>(null);
+  const [canvasReady, setCanvasReady] = useState(false);
+
+  useEffect(() => {
+    selectedCellRef.current = selectedCell;
+  }, [selectedCell]);
+
+  const updateCells = useCallback(
+    (updater: Cell[] | ((prev: Cell[]) => Cell[])) => {
+      if (typeof updater === "function") {
+        setCells((prev) => {
+          const next = updater(prev);
+          cellsRef.current = next;
+          return next;
+        });
+      } else {
+        cellsRef.current = updater;
+        setCells(updater);
+      }
+    },
+    [],
+  );
+
+  useEffect(() => {
+    const canvas = canvasRef.current;
+    if (!canvas || !canvasId || gridX === 0 || gridY === 0) {
+      return;
+    }
+
+    canvas.width = gridX * CELL_SIZE;
+    canvas.height = gridY * CELL_SIZE;
+    setCanvasReady(true);
+  }, [canvasId, gridX, gridY]);
+
+  const { viewport, updateViewport } = useCanvasViewport({
+    containerRef,
+    canvasRef,
+    gridX,
+    gridY,
+    canvasReady,
+  });
+
+  const { navigateToCoordinate } = useCanvasNavigation({
+    containerRef,
+    canvasRef,
+    updateViewport,
+  });
+
+  useCanvasRenderer({
+    canvasRef,
+    canvasReady,
+    cellsRef,
+    selectedCellRef,
+    previewColorRef,
+    votingCellIdsRef,
+    topColorMapRef,
+  });
+
+  const handleSelectCell = useCallback((cell: Cell) => {
+    setSelectedCell(cell);
+    selectedCellRef.current = cell;
+  }, []);
+
+  const { handleMouseDown, handleMouseMove, handleMouseUp, handleMouseLeave } =
+    useCanvasInteraction({
+      containerRef,
+      canvasRef,
+      cells,
+      onSelectCell: handleSelectCell,
+      onResetPreviewColor: resetPreviewColor,
+      onOpenPopup: openPopup,
+    });
+
+  const handleCanvasUpdated = useCallback(
+    ({ cellId, color }: { cellId: number; color: string }) => {
+      updateCells((prev) =>
+        prev.map((cell) =>
+          cell.id === cellId ? { ...cell, color, status: "painted" } : cell,
+        ),
+      );
+
+      if (selectedCellRef.current?.id === cellId) {
+        const nextSelectedCell: Cell = {
+          ...selectedCellRef.current,
+          color,
+          status: "painted",
+        };
+
+        setSelectedCell(nextSelectedCell);
+        selectedCellRef.current = nextSelectedCell;
+
+        if (!isRoundExpiredRef.current) {
+          closePopup();
+        }
+      }
+    },
+    [closePopup, isRoundExpiredRef, updateCells],
+  );
+
+  const clearSelectedCell = useCallback(() => {
+    setSelectedCell(null);
+    selectedCellRef.current = null;
+  }, []);
+
+  return {
+    canvasRef,
+    containerRef,
+    cells,
+    canvasId,
+    gridX,
+    gridY,
+    selectedCell,
+    viewport,
+    navigateToCoordinate,
+    setCanvasId,
+    setGridX,
+    setGridY,
+    updateCells,
+    handleMouseDown,
+    handleMouseMove,
+    handleMouseUp,
+    handleMouseLeave,
+    handleCanvasUpdated,
+    clearSelectedCell,
+  };
+}
